Move Event and associated types out of graphql_api.rs
