+++ backend-rust/src/graphql_api.rs
@@ -10,29 +10,41 @@ mod transaction_metrics;
 /// Short hand for returning API error with the message not implemented.
 macro_rules! todo_api {
     () => {
-        Err(ApiError::InternalError(String::from("Not implemented")))
+        Err(crate::graphql_api::ApiError::InternalError(String::from("Not implemented")))
     };
 }
+pub(crate) use todo_api;
 
+use crate::{
+    address::{Address, ContractAddress, ContractIndex},
+    transaction_event::{
+        baker::BakerPoolOpenStatus,
+        delegation::{BakerDelegationTarget, DelegationTarget, PassiveDelegationTarget},
+        smart_contracts::InvalidContractVersionError,
+        Event,
+    },
+    transaction_type::{
+        AccountTransaction, AccountTransactionType, CredentialDeploymentTransaction,
+        CredentialDeploymentTransactionType, DbTransactionType, TransactionType, UpdateTransaction,
+        UpdateTransactionType,
+    },
+    types::{
+        AccountIndex, Amount, BakerId, BigInteger, BlockHash, BlockHeight, DateTime, Decimal,
+        Energy, MetadataUrl, ModuleReference, TimeSpan, TransactionHash, TransactionIndex,
+    },
+};
 use account_metrics::AccountMetricsQuery;
 use anyhow::Context as _;
 use async_graphql::{
     http::GraphiQLSource,
     types::{self, connection},
-    ComplexObject, Context, EmptyMutation, Enum, InputObject, InputValueError, InputValueResult,
-    Interface, MergedObject, Object, Scalar, ScalarType, Schema, SimpleObject, Subscription, Union,
-    Value,
+    ComplexObject, Context, EmptyMutation, Enum, InputObject, Interface, MergedObject, Object,
+    Schema, SimpleObject, Subscription, Union,
 };
 use async_graphql_axum::GraphQLSubscription;
 use chrono::Duration;
 use concordium_rust_sdk::{
-    base::{
-        contracts_common::{
-            schema::{Type, VersionedModuleSchema, VersionedSchemaError},
-            Cursor,
-        },
-        smart_contracts::ReceiveName,
-    },
+    base::contracts_common::schema::{VersionedModuleSchema, VersionedSchemaError},
     id::types as sdk_types,
     types::AmountFraction,
 };
@@ -48,45 +60,6 @@ use tracing::{error, info};
 use transaction_metrics::TransactionMetricsQuery;
 
 const VERSION: &str = clap::crate_version!();
-
-struct SchemaName {
-    type_name:  &'static str,
-    value_name: &'static str,
-}
-
-enum SmartContractSchemaNames {
-    Event,
-    InputParameterReceiveFunction,
-}
-
-impl SmartContractSchemaNames {
-    pub const EVENT: SchemaName = SchemaName {
-        type_name:  "event",
-        value_name: "contract log",
-    };
-    pub const INPUT_PARAMETER_RECEIVE_FUNCTION: SchemaName = SchemaName {
-        type_name:  "receive parameter",
-        value_name: "input parameter of receive function",
-    };
-
-    pub fn value(&self) -> &'static str {
-        match self {
-            SmartContractSchemaNames::Event => Self::EVENT.value_name,
-            SmartContractSchemaNames::InputParameterReceiveFunction => {
-                Self::INPUT_PARAMETER_RECEIVE_FUNCTION.value_name
-            }
-        }
-    }
-
-    pub fn kind(&self) -> &'static str {
-        match self {
-            SmartContractSchemaNames::Event => Self::EVENT.type_name,
-            SmartContractSchemaNames::InputParameterReceiveFunction => {
-                Self::INPUT_PARAMETER_RECEIVE_FUNCTION.type_name
-            }
-        }
-    }
-}
 
 #[derive(clap::Args)]
 pub struct ApiServiceConfig {
@@ -339,7 +312,7 @@ mod monitor {
 /// return type in queries. However, some of the underlying error types are not
 /// `Clone`, so we wrap those in `Arc`s to make them `Clone`.
 #[derive(Debug, thiserror::Error, Clone)]
-enum ApiError {
+pub enum ApiError {
     #[error("Could not find resource")]
     NotFound,
     #[error("Internal error: {}", .0.message)]
@@ -370,15 +343,15 @@ impl From<sqlx::Error> for ApiError {
     fn from(value: sqlx::Error) -> Self { ApiError::FailedDatabaseQuery(Arc::new(value)) }
 }
 
-type ApiResult<A> = Result<A, ApiError>;
+pub type ApiResult<A> = Result<A, ApiError>;
 
 /// Get the database pool from the context.
-fn get_pool<'a>(ctx: &Context<'a>) -> ApiResult<&'a PgPool> {
+pub fn get_pool<'a>(ctx: &Context<'a>) -> ApiResult<&'a PgPool> {
     ctx.data::<PgPool>().map_err(ApiError::NoDatabasePool)
 }
 
 /// Get service configuration from the context.
-fn get_config<'a>(ctx: &Context<'a>) -> ApiResult<&'a ApiServiceConfig> {
+pub fn get_config<'a>(ctx: &Context<'a>) -> ApiResult<&'a ApiServiceConfig> {
     ctx.data::<ApiServiceConfig>().map_err(ApiError::NoServiceConfig)
 }
 
@@ -441,7 +414,7 @@ impl<A> ConnectionQuery<A> {
 }
 
 #[derive(Default)]
-pub struct BaseQuery;
+struct BaseQuery;
 
 #[Object]
 #[allow(clippy::too_many_arguments)]
@@ -1089,8 +1062,8 @@ LIMIT 30", // WHERE slot_time > (LOCALTIMESTAMP - $1::interval)
 }
 
 pub struct Subscription {
-    pub block_added:      broadcast::Receiver<Block>,
-    pub accounts_updated: broadcast::Receiver<AccountsUpdatedSubscriptionItem>,
+    block_added:      broadcast::Receiver<Block>,
+    accounts_updated: broadcast::Receiver<AccountsUpdatedSubscriptionItem>,
 }
 
 impl Subscription {
@@ -1241,163 +1214,13 @@ pub struct AccountsUpdatedSubscriptionItem {
     address: String,
 }
 
-/// The UnsignedLong scalar type represents a unsigned 64-bit numeric
-/// non-fractional value greater than or equal to 0.
-#[derive(
-    Clone,
-    Copy,
-    derive_more::Display,
-    Debug,
-    serde::Serialize,
-    serde::Deserialize,
-    derive_more::From,
-    derive_more::FromStr,
-)]
-#[repr(transparent)]
-#[serde(transparent)]
-struct UnsignedLong(u64);
-#[Scalar]
-impl ScalarType for UnsignedLong {
-    fn parse(value: Value) -> InputValueResult<Self> {
-        let Value::Number(number) = &value else {
-            return Err(InputValueError::expected_type(value));
-        };
-        if let Some(v) = number.as_u64() {
-            Ok(Self(v))
-        } else {
-            Err(InputValueError::expected_type(value))
-        }
-    }
-
-    fn to_value(&self) -> Value { Value::Number(self.0.into()) }
-}
-
-impl TryFrom<i64> for UnsignedLong {
-    type Error = <u64 as TryFrom<i64>>::Error;
-
-    fn try_from(number: i64) -> Result<Self, Self::Error> { Ok(UnsignedLong(number.try_into()?)) }
-}
-
-/// The `Long` scalar type represents non-fractional signed whole 64-bit numeric
-/// values. Long can represent values between -(2^63) and 2^63 - 1.
-#[derive(serde::Serialize, serde::Deserialize, derive_more::From)]
-#[repr(transparent)]
-#[serde(transparent)]
-struct Long(i64);
-#[Scalar]
-impl ScalarType for Long {
-    fn parse(value: Value) -> InputValueResult<Self> {
-        let Value::Number(number) = &value else {
-            return Err(InputValueError::expected_type(value));
-        };
-        if let Some(v) = number.as_i64() {
-            Ok(Self(v))
-        } else {
-            Err(InputValueError::expected_type(value))
-        }
-    }
-
-    fn to_value(&self) -> Value { Value::Number(self.0.into()) }
-}
-#[derive(serde::Serialize, serde::Deserialize, derive_more::From)]
-#[repr(transparent)]
-#[serde(transparent)]
-struct Byte(u8);
-#[Scalar]
-impl ScalarType for Byte {
-    fn parse(value: Value) -> InputValueResult<Self> {
-        let Value::Number(number) = &value else {
-            return Err(InputValueError::expected_type(value));
-        };
-        let Some(v) = number.as_u64() else {
-            return Err(InputValueError::expected_type(value));
-        };
-
-        if let Ok(v) = u8::try_from(v) {
-            Ok(Self(v))
-        } else {
-            Err(InputValueError::expected_type(value))
-        }
-    }
-
-    fn to_value(&self) -> Value { Value::Number(self.0.into()) }
-}
-
-#[derive(serde::Serialize, serde::Deserialize, derive_more::From)]
-#[repr(transparent)]
-#[serde(transparent)]
-struct Decimal(rust_decimal::Decimal);
-#[Scalar]
-impl ScalarType for Decimal {
-    fn parse(value: Value) -> InputValueResult<Self> {
-        let Value::String(string) = value else {
-            return Err(InputValueError::expected_type(value));
-        };
-        Ok(Self(rust_decimal::Decimal::from_str(string.as_str())?))
-    }
-
-    fn to_value(&self) -> Value { Value::String(self.0.to_string()) }
-}
-
-impl From<concordium_rust_sdk::types::AmountFraction> for Decimal {
-    fn from(fraction: concordium_rust_sdk::types::AmountFraction) -> Self {
-        Self(concordium_rust_sdk::types::PartsPerHundredThousands::from(fraction).into())
-    }
-}
-
-/// The `TimeSpan` scalar represents an ISO-8601 compliant duration type.
-#[derive(serde::Serialize, serde::Deserialize, Clone)]
-#[repr(transparent)]
-#[serde(try_from = "String", into = "String")]
-struct TimeSpan(Duration);
-#[Scalar]
-impl ScalarType for TimeSpan {
-    fn parse(value: Value) -> InputValueResult<Self> {
-        let Value::String(string) = value else {
-            return Err(InputValueError::expected_type(value));
-        };
-        Ok(Self::try_from(string)?)
-    }
-
-    fn to_value(&self) -> Value { Value::String(self.0.to_string()) }
-}
-impl TryFrom<String> for TimeSpan {
-    type Error = anyhow::Error;
-
-    fn try_from(value: String) -> Result<Self, Self::Error> {
-        let duration = iso8601_duration::Duration::from_str(value.as_str())
-            .map_err(|err| anyhow::anyhow!("Invalid duration, expected ISO-8601"))?;
-        Ok(Self(duration.to_chrono().context("Failed to construct duration")?))
-    }
-}
-impl From<TimeSpan> for String {
-    fn from(time: TimeSpan) -> Self { time.0.to_string() }
-}
-impl From<Duration> for TimeSpan {
-    fn from(duration: Duration) -> Self { TimeSpan(duration) }
-}
-
-type BlockHeight = i64;
-type BlockHash = String;
-type TransactionHash = String;
-type ModuleReference = String;
-type BakerId = i64;
-type AccountIndex = i64;
-type TransactionIndex = i64;
-type Amount = i64; // TODO: should be UnsignedLong in graphQL
-type Energy = i64; // TODO: should be UnsignedLong in graphQL
-type DateTime = chrono::DateTime<chrono::Utc>; // TODO check format matches.
-type ContractIndex = UnsignedLong; // TODO check format.
-type BigInteger = u64; // TODO check format.
-type MetadataUrl = String;
-
 #[derive(SimpleObject)]
 struct Versions {
     backend_versions: String,
 }
 
 #[derive(Debug, Clone, sqlx::FromRow)]
-pub struct Block {
+struct Block {
     hash:              BlockHash,
     height:            BlockHeight,
     /// Time of the block being baked.
@@ -2588,22 +2411,7 @@ struct PassiveDelegationPoolRewardTarget {
 }
 
 #[derive(SimpleObject, serde::Serialize, serde::Deserialize)]
-struct PassiveDelegationTarget {
-    #[graphql(
-        name = "_",
-        deprecation = "Don't use! This field is only in the schema to make this a valid GraphQL \
-                       type (which does not allow types without any fields)"
-    )]
-    dummy: bool,
-}
-
-#[derive(SimpleObject, serde::Serialize, serde::Deserialize)]
 struct BakerPoolRewardTarget {
-    baker_id: BakerId,
-}
-
-#[derive(SimpleObject, serde::Serialize, serde::Deserialize)]
-struct BakerDelegationTarget {
     baker_id: BakerId,
 }
 
@@ -2905,168 +2713,6 @@ impl Transaction {
             Ok(TransactionResult::Rejected(Rejected {
                 reason,
             }))
-        }
-    }
-}
-
-#[derive(Union)]
-#[allow(clippy::enum_variant_names)]
-enum TransactionType {
-    AccountTransaction(AccountTransaction),
-    CredentialDeploymentTransaction(CredentialDeploymentTransaction),
-    UpdateTransaction(UpdateTransaction),
-}
-
-#[derive(SimpleObject)]
-struct AccountTransaction {
-    account_transaction_type: Option<AccountTransactionType>,
-}
-
-#[derive(Enum, Clone, Copy, PartialEq, Eq, sqlx::Type)]
-#[sqlx(type_name = "account_transaction_type")]
-pub enum AccountTransactionType {
-    InitializeSmartContractInstance,
-    UpdateSmartContractInstance,
-    SimpleTransfer,
-    EncryptedTransfer,
-    SimpleTransferWithMemo,
-    EncryptedTransferWithMemo,
-    TransferWithScheduleWithMemo,
-    DeployModule,
-    AddBaker,
-    RemoveBaker,
-    UpdateBakerStake,
-    UpdateBakerRestakeEarnings,
-    UpdateBakerKeys,
-    UpdateCredentialKeys,
-    TransferToEncrypted,
-    TransferToPublic,
-    TransferWithSchedule,
-    UpdateCredentials,
-    RegisterData,
-    ConfigureBaker,
-    ConfigureDelegation,
-}
-
-impl From<concordium_rust_sdk::types::TransactionType> for AccountTransactionType {
-    fn from(value: concordium_rust_sdk::types::TransactionType) -> Self {
-        use concordium_rust_sdk::types::TransactionType as TT;
-        use AccountTransactionType as ATT;
-        #[allow(deprecated)]
-        match value {
-            TT::DeployModule => ATT::DeployModule,
-            TT::InitContract => ATT::InitializeSmartContractInstance,
-            TT::Update => ATT::UpdateSmartContractInstance,
-            TT::Transfer => ATT::SimpleTransfer,
-            TT::AddBaker => ATT::AddBaker,
-            TT::RemoveBaker => ATT::RemoveBaker,
-            TT::UpdateBakerStake => ATT::UpdateBakerStake,
-            TT::UpdateBakerRestakeEarnings => ATT::UpdateBakerRestakeEarnings,
-            TT::UpdateBakerKeys => ATT::UpdateBakerKeys,
-            TT::UpdateCredentialKeys => ATT::UpdateCredentialKeys,
-            TT::EncryptedAmountTransfer => ATT::EncryptedTransfer,
-            TT::TransferToEncrypted => ATT::TransferToEncrypted,
-            TT::TransferToPublic => ATT::TransferToPublic,
-            TT::TransferWithSchedule => ATT::TransferWithSchedule,
-            TT::UpdateCredentials => ATT::UpdateCredentials,
-            TT::RegisterData => ATT::RegisterData,
-            TT::TransferWithMemo => ATT::SimpleTransferWithMemo,
-            TT::EncryptedAmountTransferWithMemo => ATT::EncryptedTransferWithMemo,
-            TT::TransferWithScheduleAndMemo => ATT::TransferWithScheduleWithMemo,
-            TT::ConfigureBaker => ATT::ConfigureBaker,
-            TT::ConfigureDelegation => ATT::ConfigureDelegation,
-        }
-    }
-}
-
-#[derive(SimpleObject)]
-struct CredentialDeploymentTransaction {
-    credential_deployment_transaction_type: CredentialDeploymentTransactionType,
-}
-
-#[derive(Enum, Clone, Copy, PartialEq, Eq, sqlx::Type)]
-#[sqlx(type_name = "credential_deployment_transaction_type")]
-pub enum CredentialDeploymentTransactionType {
-    Initial,
-    Normal,
-}
-
-impl From<concordium_rust_sdk::types::CredentialType> for CredentialDeploymentTransactionType {
-    fn from(value: concordium_rust_sdk::types::CredentialType) -> Self {
-        use concordium_rust_sdk::types::CredentialType;
-        match value {
-            CredentialType::Initial => CredentialDeploymentTransactionType::Initial,
-            CredentialType::Normal => CredentialDeploymentTransactionType::Normal,
-        }
-    }
-}
-
-#[derive(SimpleObject)]
-struct UpdateTransaction {
-    update_transaction_type: UpdateTransactionType,
-}
-
-#[derive(Enum, Clone, Copy, PartialEq, Eq, sqlx::Type)]
-#[sqlx(type_name = "update_transaction_type")]
-pub enum UpdateTransactionType {
-    UpdateProtocol,
-    UpdateElectionDifficulty,
-    UpdateEuroPerEnergy,
-    UpdateMicroGtuPerEuro,
-    UpdateFoundationAccount,
-    UpdateMintDistribution,
-    UpdateTransactionFeeDistribution,
-    UpdateGasRewards,
-    UpdateBakerStakeThreshold,
-    UpdateAddAnonymityRevoker,
-    UpdateAddIdentityProvider,
-    UpdateRootKeys,
-    UpdateLevel1Keys,
-    UpdateLevel2Keys,
-    UpdatePoolParameters,
-    UpdateCooldownParameters,
-    UpdateTimeParameters,
-    MintDistributionCpv1Update,
-    GasRewardsCpv2Update,
-    TimeoutParametersUpdate,
-    MinBlockTimeUpdate,
-    BlockEnergyLimitUpdate,
-    FinalizationCommitteeParametersUpdate,
-}
-
-impl From<concordium_rust_sdk::types::UpdateType> for UpdateTransactionType {
-    fn from(value: concordium_rust_sdk::types::UpdateType) -> Self {
-        use concordium_rust_sdk::types::UpdateType;
-        match value {
-            UpdateType::UpdateProtocol => UpdateTransactionType::UpdateProtocol,
-            UpdateType::UpdateElectionDifficulty => UpdateTransactionType::UpdateElectionDifficulty,
-            UpdateType::UpdateEuroPerEnergy => UpdateTransactionType::UpdateEuroPerEnergy,
-            UpdateType::UpdateMicroGTUPerEuro => UpdateTransactionType::UpdateMicroGtuPerEuro,
-            UpdateType::UpdateFoundationAccount => UpdateTransactionType::UpdateFoundationAccount,
-            UpdateType::UpdateMintDistribution => UpdateTransactionType::UpdateMintDistribution,
-            UpdateType::UpdateTransactionFeeDistribution => {
-                UpdateTransactionType::UpdateTransactionFeeDistribution
-            }
-            UpdateType::UpdateGASRewards => UpdateTransactionType::UpdateGasRewards,
-            UpdateType::UpdateAddAnonymityRevoker => {
-                UpdateTransactionType::UpdateAddAnonymityRevoker
-            }
-            UpdateType::UpdateAddIdentityProvider => {
-                UpdateTransactionType::UpdateAddIdentityProvider
-            }
-            UpdateType::UpdateRootKeys => UpdateTransactionType::UpdateRootKeys,
-            UpdateType::UpdateLevel1Keys => UpdateTransactionType::UpdateLevel1Keys,
-            UpdateType::UpdateLevel2Keys => UpdateTransactionType::UpdateLevel2Keys,
-            UpdateType::UpdatePoolParameters => UpdateTransactionType::UpdatePoolParameters,
-            UpdateType::UpdateCooldownParameters => UpdateTransactionType::UpdateCooldownParameters,
-            UpdateType::UpdateTimeParameters => UpdateTransactionType::UpdateTimeParameters,
-            UpdateType::UpdateGASRewardsCPV2 => UpdateTransactionType::GasRewardsCpv2Update,
-            UpdateType::UpdateTimeoutParameters => UpdateTransactionType::TimeoutParametersUpdate,
-            UpdateType::UpdateMinBlockTime => UpdateTransactionType::MinBlockTimeUpdate,
-            UpdateType::UpdateBlockEnergyLimit => UpdateTransactionType::BlockEnergyLimitUpdate,
-            UpdateType::UpdateFinalizationCommitteeParameters => {
-                UpdateTransactionType::FinalizationCommitteeParametersUpdate
-            }
         }
     }
 }
@@ -3743,25 +3389,6 @@ struct CommissionRates {
     baking_commission:       Option<Decimal>,
 }
 
-#[derive(Enum, Copy, Clone, PartialEq, Eq, serde::Serialize, serde::Deserialize, sqlx::Type)]
-#[sqlx(type_name = "pool_open_status")] // only for PostgreSQL to match a type definition
-pub enum BakerPoolOpenStatus {
-    OpenForAll,
-    ClosedForNew,
-    ClosedForAll,
-}
-
-impl From<concordium_rust_sdk::types::OpenStatus> for BakerPoolOpenStatus {
-    fn from(status: concordium_rust_sdk::types::OpenStatus) -> Self {
-        use concordium_rust_sdk::types::OpenStatus;
-        match status {
-            OpenStatus::OpenForAll => Self::OpenForAll,
-            OpenStatus::ClosedForNew => Self::ClosedForNew,
-            OpenStatus::ClosedForAll => Self::ClosedForAll,
-        }
-    }
-}
-
 #[derive(SimpleObject)]
 struct Ranking {
     rank:  i32,
@@ -3774,32 +3401,6 @@ struct Delegation {
     staked_amount:     Amount,
     restake_earnings:  bool,
     delegation_target: DelegationTarget,
-}
-
-#[derive(Union, serde::Serialize, serde::Deserialize)]
-enum DelegationTarget {
-    PassiveDelegationTarget(PassiveDelegationTarget),
-    BakerDelegationTarget(BakerDelegationTarget),
-}
-
-impl TryFrom<concordium_rust_sdk::types::DelegationTarget> for DelegationTarget {
-    type Error = anyhow::Error;
-
-    fn try_from(target: concordium_rust_sdk::types::DelegationTarget) -> Result<Self, Self::Error> {
-        use concordium_rust_sdk::types::DelegationTarget as Target;
-        match target {
-            Target::Passive => {
-                Ok(DelegationTarget::PassiveDelegationTarget(PassiveDelegationTarget {
-                    dummy: true,
-                }))
-            }
-            Target::Baker {
-                baker_id,
-            } => Ok(DelegationTarget::BakerDelegationTarget(BakerDelegationTarget {
-                baker_id: baker_id.id.index.try_into()?,
-            })),
-        }
-    }
 }
 
 #[derive(Union)]
@@ -4059,609 +3660,6 @@ impl SearchResult {
     ) -> ApiResult<connection::Connection<String, NodeStatus>> {
         todo_api!()
     }
-}
-
-fn decode_value_with_schema(
-    opt_schema: Option<&Type>,
-    value: &[u8],
-    schema_name: SmartContractSchemaNames,
-) -> String {
-    let Some(schema) = opt_schema else {
-        // Note: There could be something better displayed than this string if no schema
-        // is available for decoding at the frontend long-term.
-        return format!(
-            "No embedded {} schema in smart contract available for decoding",
-            schema_name.kind()
-        );
-    };
-
-    let mut cursor = Cursor::new(&value);
-    match schema.to_json(&mut cursor) {
-        Ok(v) => {
-            serde_json::to_string(&v).unwrap_or_else(|e| {
-                // We don't return an error here since the query is correctly formed and
-                // the CCDScan backend is working as expected.
-                // A wrong/missing schema is a mistake by the smart contract
-                // developer which in general cannot be fixed after the deployment of
-                // the contract. We display the error message (instead of the decoded
-                // value) in the block explorer to make the info visible to the smart
-                // contract developer for debugging purposes here.
-                format!(
-                    "Failed to deserialize {} with {} schema into string: {:?}",
-                    schema_name.value(),
-                    schema_name.kind(),
-                    e
-                )
-            })
-        }
-        Err(e) => {
-            // We don't return an error here since the query is correctly formed and
-            // the CCDScan backend is working as expected.
-            // A wrong/missing schema is a mistake by the smart contract
-            // developer which in general cannot be fixed after the deployment of
-            // the contract. We display the error message (instead of the decoded
-            // value) in the block explorer to make the info visible to the smart
-            // contract developer for debugging purposes here.
-            format!(
-                "Failed to deserialize {} with {} schema: {:?}",
-                schema_name.value(),
-                schema_name.kind(),
-                e.display(true)
-            )
-        }
-    }
-}
-
-#[derive(SimpleObject, serde::Serialize, serde::Deserialize)]
-#[graphql(complex)]
-struct ContractAddress {
-    index:     ContractIndex,
-    sub_index: ContractIndex,
-}
-#[ComplexObject]
-impl ContractAddress {
-    async fn as_string(&self) -> String {
-        concordium_rust_sdk::types::ContractAddress::new(self.index.0, self.sub_index.0).to_string()
-    }
-}
-impl ContractAddress {
-    fn new(index: i64, sub_index: i64) -> ApiResult<Self> {
-        Ok(Self {
-            index:     u64::try_from(index)?.into(),
-            sub_index: u64::try_from(index)?.into(),
-        })
-    }
-}
-
-impl From<concordium_rust_sdk::types::ContractAddress> for ContractAddress {
-    fn from(value: concordium_rust_sdk::types::ContractAddress) -> Self {
-        Self {
-            index:     value.index.into(),
-            sub_index: value.subindex.into(),
-        }
-    }
-}
-
-#[derive(Union, serde::Serialize, serde::Deserialize)]
-enum Address {
-    ContractAddress(ContractAddress),
-    AccountAddress(AccountAddress),
-}
-
-impl From<concordium_rust_sdk::types::Address> for Address {
-    fn from(value: concordium_rust_sdk::types::Address) -> Self {
-        use concordium_rust_sdk::types::Address as Addr;
-        match value {
-            Addr::Account(a) => Address::AccountAddress(a.into()),
-            Addr::Contract(c) => Address::ContractAddress(c.into()),
-        }
-    }
-}
-
-#[derive(Union, serde::Serialize, serde::Deserialize)]
-pub enum Event {
-    /// A transfer of CCD. Can be either from an account or a smart contract
-    /// instance, but the receiver in this event is always an account.
-    Transferred(Transferred),
-    AccountCreated(AccountCreated),
-    AmountAddedByDecryption(AmountAddedByDecryption),
-    BakerAdded(BakerAdded),
-    BakerKeysUpdated(BakerKeysUpdated),
-    BakerRemoved(BakerRemoved),
-    BakerSetRestakeEarnings(BakerSetRestakeEarnings),
-    BakerStakeDecreased(BakerStakeDecreased),
-    BakerStakeIncreased(BakerStakeIncreased),
-    ContractInitialized(ContractInitialized),
-    ContractModuleDeployed(ContractModuleDeployed),
-    ContractUpdated(ContractUpdated),
-    ContractCall(ContractCall),
-    CredentialDeployed(CredentialDeployed),
-    CredentialKeysUpdated(CredentialKeysUpdated),
-    CredentialsUpdated(CredentialsUpdated),
-    DataRegistered(DataRegistered),
-    EncryptedAmountsRemoved(EncryptedAmountsRemoved),
-    EncryptedSelfAmountAdded(EncryptedSelfAmountAdded),
-    NewEncryptedAmount(NewEncryptedAmount),
-    TransferMemo(TransferMemo),
-    TransferredWithSchedule(TransferredWithSchedule),
-    ChainUpdateEnqueued(ChainUpdateEnqueued),
-    ContractInterrupted(ContractInterrupted),
-    ContractResumed(ContractResumed),
-    ContractUpgraded(ContractUpgraded),
-    BakerSetOpenStatus(BakerSetOpenStatus),
-    BakerSetMetadataURL(BakerSetMetadataURL),
-    BakerSetTransactionFeeCommission(BakerSetTransactionFeeCommission),
-    BakerSetBakingRewardCommission(BakerSetBakingRewardCommission),
-    BakerSetFinalizationRewardCommission(BakerSetFinalizationRewardCommission),
-    DelegationAdded(DelegationAdded),
-    DelegationRemoved(DelegationRemoved),
-    DelegationStakeIncreased(DelegationStakeIncreased),
-    DelegationStakeDecreased(DelegationStakeDecreased),
-    DelegationSetRestakeEarnings(DelegationSetRestakeEarnings),
-    DelegationSetDelegationTarget(DelegationSetDelegationTarget),
-}
-
-pub fn events_from_summary(
-    value: concordium_rust_sdk::types::BlockItemSummaryDetails,
-) -> anyhow::Result<Vec<Event>> {
-    use concordium_rust_sdk::types::{AccountTransactionEffects, BlockItemSummaryDetails};
-    let events = match value {
-        BlockItemSummaryDetails::AccountTransaction(details) => match details.effects {
-            AccountTransactionEffects::None {
-                ..
-            } => {
-                anyhow::bail!("Transaction was rejected")
-            }
-            AccountTransactionEffects::ModuleDeployed {
-                module_ref,
-            } => {
-                vec![Event::ContractModuleDeployed(ContractModuleDeployed {
-                    module_ref: module_ref.to_string(),
-                })]
-            }
-            AccountTransactionEffects::ContractInitialized {
-                data,
-            } => {
-                vec![Event::ContractInitialized(ContractInitialized {
-                    module_ref:        data.origin_ref.to_string(),
-                    contract_address:  data.address.into(),
-                    amount:            i64::try_from(data.amount.micro_ccd)?,
-                    init_name:         data.init_name.to_string(),
-                    version:           data.contract_version.into(),
-                    contract_logs_raw: data.events.iter().map(|e| e.as_ref().to_vec()).collect(),
-                })]
-            }
-            AccountTransactionEffects::ContractUpdateIssued {
-                effects,
-            } => {
-                use concordium_rust_sdk::types::ContractTraceElement;
-                effects
-                    .into_iter()
-                    .map(|effect| match effect {
-                        ContractTraceElement::Updated {
-                            data,
-                        } => Ok(Event::ContractUpdated(ContractUpdated {
-                            contract_address:  data.address.into(),
-                            instigator:        data.instigator.into(),
-                            amount:            data.amount.micro_ccd().try_into()?,
-                            receive_name:      data.receive_name.to_string(),
-                            version:           data.contract_version.into(),
-                            input_parameter:   data.message.as_ref().to_vec(),
-                            contract_logs_raw: data
-                                .events
-                                .iter()
-                                .map(|e| e.as_ref().to_vec())
-                                .collect(),
-                        })),
-                        ContractTraceElement::Transferred {
-                            from,
-                            amount,
-                            to,
-                        } => Ok(Event::Transferred(Transferred {
-                            amount: amount.micro_ccd().try_into()?,
-                            from:   Address::ContractAddress(from.into()),
-                            to:     to.into(),
-                        })),
-                        ContractTraceElement::Interrupted {
-                            address,
-                            events,
-                        } => Ok(Event::ContractInterrupted(ContractInterrupted {
-                            contract_address:  address.into(),
-                            contract_logs_raw: events.iter().map(|e| e.as_ref().to_vec()).collect(),
-                        })),
-                        ContractTraceElement::Resumed {
-                            address,
-                            success,
-                        } => Ok(Event::ContractResumed(ContractResumed {
-                            contract_address: address.into(),
-                            success,
-                        })),
-                        ContractTraceElement::Upgraded {
-                            address,
-                            from,
-                            to,
-                        } => Ok(Event::ContractUpgraded(ContractUpgraded {
-                            contract_address: address.into(),
-                            from:             from.to_string(),
-                            to:               to.to_string(),
-                        })),
-                    })
-                    .collect::<anyhow::Result<Vec<_>>>()?
-            }
-            AccountTransactionEffects::AccountTransfer {
-                amount,
-                to,
-            } => {
-                vec![Event::Transferred(Transferred {
-                    amount: i64::try_from(amount.micro_ccd)?,
-                    from:   Address::AccountAddress(details.sender.into()),
-                    to:     to.into(),
-                })]
-            }
-            AccountTransactionEffects::AccountTransferWithMemo {
-                amount,
-                to,
-                memo,
-            } => {
-                vec![
-                    Event::Transferred(Transferred {
-                        amount: i64::try_from(amount.micro_ccd)?,
-                        from:   Address::AccountAddress(details.sender.into()),
-                        to:     to.into(),
-                    }),
-                    Event::TransferMemo(memo.into()),
-                ]
-            }
-            AccountTransactionEffects::BakerAdded {
-                data,
-            } => {
-                vec![Event::BakerAdded(BakerAdded {
-                    staked_amount:    data.stake.micro_ccd.try_into()?,
-                    restake_earnings: data.restake_earnings,
-                    baker_id:         data.keys_event.baker_id.id.index.try_into()?,
-                    sign_key:         serde_json::to_string(&data.keys_event.sign_key)?,
-                    election_key:     serde_json::to_string(&data.keys_event.election_key)?,
-                    aggregation_key:  serde_json::to_string(&data.keys_event.aggregation_key)?,
-                })]
-            }
-            AccountTransactionEffects::BakerRemoved {
-                baker_id,
-            } => {
-                vec![Event::BakerRemoved(BakerRemoved {
-                    baker_id: baker_id.id.index.try_into()?,
-                })]
-            }
-            AccountTransactionEffects::BakerStakeUpdated {
-                data,
-            } => {
-                if let Some(data) = data {
-                    if data.increased {
-                        vec![Event::BakerStakeIncreased(BakerStakeIncreased {
-                            baker_id:          data.baker_id.id.index.try_into()?,
-                            new_staked_amount: data.new_stake.micro_ccd.try_into()?,
-                        })]
-                    } else {
-                        vec![Event::BakerStakeDecreased(BakerStakeDecreased {
-                            baker_id:          data.baker_id.id.index.try_into()?,
-                            new_staked_amount: data.new_stake.micro_ccd.try_into()?,
-                        })]
-                    }
-                } else {
-                    Vec::new()
-                }
-            }
-            AccountTransactionEffects::BakerRestakeEarningsUpdated {
-                baker_id,
-                restake_earnings,
-            } => {
-                vec![Event::BakerSetRestakeEarnings(BakerSetRestakeEarnings {
-                    baker_id: baker_id.id.index.try_into()?,
-                    restake_earnings,
-                })]
-            }
-            AccountTransactionEffects::BakerKeysUpdated {
-                data,
-            } => {
-                vec![Event::BakerKeysUpdated(BakerKeysUpdated {
-                    baker_id:        data.baker_id.id.index.try_into()?,
-                    sign_key:        serde_json::to_string(&data.sign_key)?,
-                    election_key:    serde_json::to_string(&data.election_key)?,
-                    aggregation_key: serde_json::to_string(&data.aggregation_key)?,
-                })]
-            }
-            AccountTransactionEffects::EncryptedAmountTransferred {
-                removed,
-                added,
-            } => {
-                vec![
-                    Event::EncryptedAmountsRemoved((*removed).try_into()?),
-                    Event::NewEncryptedAmount((*added).try_into()?),
-                ]
-            }
-            AccountTransactionEffects::EncryptedAmountTransferredWithMemo {
-                removed,
-                added,
-                memo,
-            } => {
-                vec![
-                    Event::EncryptedAmountsRemoved((*removed).try_into()?),
-                    Event::NewEncryptedAmount((*added).try_into()?),
-                    Event::TransferMemo(memo.into()),
-                ]
-            }
-            AccountTransactionEffects::TransferredToEncrypted {
-                data,
-            } => {
-                vec![Event::EncryptedSelfAmountAdded(EncryptedSelfAmountAdded {
-                    account_address:      data.account.into(),
-                    new_encrypted_amount: serde_json::to_string(&data.new_amount)?,
-                    amount:               data.amount.micro_ccd.try_into()?,
-                })]
-            }
-            AccountTransactionEffects::TransferredToPublic {
-                removed,
-                amount,
-            } => {
-                vec![
-                    Event::EncryptedAmountsRemoved((*removed).try_into()?),
-                    Event::AmountAddedByDecryption(AmountAddedByDecryption {
-                        amount:          amount.micro_ccd().try_into()?,
-                        account_address: details.sender.into(),
-                    }),
-                ]
-            }
-            AccountTransactionEffects::TransferredWithSchedule {
-                to,
-                amount,
-            } => {
-                vec![Event::TransferredWithSchedule(TransferredWithSchedule {
-                    from_account_address: details.sender.into(),
-                    to_account_address:   to.into(),
-                    total_amount:         amount
-                        .into_iter()
-                        .map(|(_, amount)| amount.micro_ccd())
-                        .sum::<u64>()
-                        .try_into()?,
-                })]
-            }
-            AccountTransactionEffects::TransferredWithScheduleAndMemo {
-                to,
-                amount,
-                memo,
-            } => {
-                vec![
-                    Event::TransferredWithSchedule(TransferredWithSchedule {
-                        from_account_address: details.sender.into(),
-                        to_account_address:   to.into(),
-                        total_amount:         amount
-                            .into_iter()
-                            .map(|(_, amount)| amount.micro_ccd())
-                            .sum::<u64>()
-                            .try_into()?,
-                    }),
-                    Event::TransferMemo(memo.into()),
-                ]
-            }
-            AccountTransactionEffects::CredentialKeysUpdated {
-                cred_id,
-            } => {
-                vec![Event::CredentialKeysUpdated(CredentialKeysUpdated {
-                    cred_id: cred_id.to_string(),
-                })]
-            }
-            AccountTransactionEffects::CredentialsUpdated {
-                new_cred_ids,
-                removed_cred_ids,
-                new_threshold,
-            } => {
-                vec![Event::CredentialsUpdated(CredentialsUpdated {
-                    account_address:  details.sender.into(),
-                    new_cred_ids:     new_cred_ids
-                        .into_iter()
-                        .map(|cred| cred.to_string())
-                        .collect(),
-                    removed_cred_ids: removed_cred_ids
-                        .into_iter()
-                        .map(|cred| cred.to_string())
-                        .collect(),
-                    new_threshold:    Byte(u8::from(new_threshold)),
-                })]
-            }
-            AccountTransactionEffects::DataRegistered {
-                data,
-            } => {
-                vec![Event::DataRegistered(DataRegistered {
-                    data_as_hex: hex::encode(data.as_ref()),
-                })]
-            }
-            AccountTransactionEffects::BakerConfigured {
-                data,
-            } => data
-                .into_iter()
-                .map(|baker_event| {
-                    use concordium_rust_sdk::types::BakerEvent;
-                    match baker_event {
-                        BakerEvent::BakerAdded {
-                            data,
-                        } => Ok(Event::BakerAdded(BakerAdded {
-                            staked_amount:    data.stake.micro_ccd.try_into()?,
-                            restake_earnings: data.restake_earnings,
-                            baker_id:         data.keys_event.baker_id.id.index.try_into()?,
-                            sign_key:         serde_json::to_string(&data.keys_event.sign_key)?,
-                            election_key:     serde_json::to_string(&data.keys_event.election_key)?,
-                            aggregation_key:  serde_json::to_string(
-                                &data.keys_event.aggregation_key,
-                            )?,
-                        })),
-                        BakerEvent::BakerRemoved {
-                            baker_id,
-                        } => Ok(Event::BakerRemoved(BakerRemoved {
-                            baker_id: baker_id.id.index.try_into()?,
-                        })),
-                        BakerEvent::BakerStakeIncreased {
-                            baker_id,
-                            new_stake,
-                        } => Ok(Event::BakerStakeIncreased(BakerStakeIncreased {
-                            baker_id:          baker_id.id.index.try_into()?,
-                            new_staked_amount: new_stake.micro_ccd.try_into()?,
-                        })),
-                        BakerEvent::BakerStakeDecreased {
-                            baker_id,
-                            new_stake,
-                        } => Ok(Event::BakerStakeDecreased(BakerStakeDecreased {
-                            baker_id:          baker_id.id.index.try_into()?,
-                            new_staked_amount: new_stake.micro_ccd.try_into()?,
-                        })),
-                        BakerEvent::BakerRestakeEarningsUpdated {
-                            baker_id,
-                            restake_earnings,
-                        } => Ok(Event::BakerSetRestakeEarnings(BakerSetRestakeEarnings {
-                            baker_id: baker_id.id.index.try_into()?,
-                            restake_earnings,
-                        })),
-                        BakerEvent::BakerKeysUpdated {
-                            data,
-                        } => Ok(Event::BakerKeysUpdated(BakerKeysUpdated {
-                            baker_id:        data.baker_id.id.index.try_into()?,
-                            sign_key:        serde_json::to_string(&data.sign_key)?,
-                            election_key:    serde_json::to_string(&data.election_key)?,
-                            aggregation_key: serde_json::to_string(&data.aggregation_key)?,
-                        })),
-                        BakerEvent::BakerSetOpenStatus {
-                            baker_id,
-                            open_status,
-                        } => Ok(Event::BakerSetOpenStatus(BakerSetOpenStatus {
-                            baker_id:        baker_id.id.index.try_into()?,
-                            account_address: details.sender.into(),
-                            open_status:     open_status.into(),
-                        })),
-                        BakerEvent::BakerSetMetadataURL {
-                            baker_id,
-                            metadata_url,
-                        } => Ok(Event::BakerSetMetadataURL(BakerSetMetadataURL {
-                            baker_id:        baker_id.id.index.try_into()?,
-                            account_address: details.sender.into(),
-                            metadata_url:    metadata_url.into(),
-                        })),
-                        BakerEvent::BakerSetTransactionFeeCommission {
-                            baker_id,
-                            transaction_fee_commission,
-                        } => Ok(Event::BakerSetTransactionFeeCommission(
-                            BakerSetTransactionFeeCommission {
-                                baker_id:                   baker_id.id.index.try_into()?,
-                                account_address:            details.sender.into(),
-                                transaction_fee_commission: transaction_fee_commission.into(),
-                            },
-                        )),
-                        BakerEvent::BakerSetBakingRewardCommission {
-                            baker_id,
-                            baking_reward_commission,
-                        } => Ok(Event::BakerSetBakingRewardCommission(
-                            BakerSetBakingRewardCommission {
-                                baker_id:                 baker_id.id.index.try_into()?,
-                                account_address:          details.sender.into(),
-                                baking_reward_commission: baking_reward_commission.into(),
-                            },
-                        )),
-                        BakerEvent::BakerSetFinalizationRewardCommission {
-                            baker_id,
-                            finalization_reward_commission,
-                        } => Ok(Event::BakerSetFinalizationRewardCommission(
-                            BakerSetFinalizationRewardCommission {
-                                baker_id: baker_id.id.index.try_into()?,
-                                account_address: details.sender.into(),
-                                finalization_reward_commission: finalization_reward_commission
-                                    .into(),
-                            },
-                        )),
-                        BakerEvent::DelegationRemoved {
-                            delegator_id,
-                        } => {
-                            unimplemented!()
-                        }
-                    }
-                })
-                .collect::<anyhow::Result<Vec<Event>>>()?,
-            AccountTransactionEffects::DelegationConfigured {
-                data,
-            } => {
-                use concordium_rust_sdk::types::DelegationEvent;
-                data.into_iter()
-                    .map(|event| match event {
-                        DelegationEvent::DelegationStakeIncreased {
-                            delegator_id,
-                            new_stake,
-                        } => Ok(Event::DelegationStakeIncreased(DelegationStakeIncreased {
-                            delegator_id:      delegator_id.id.index.try_into()?,
-                            account_address:   details.sender.into(),
-                            new_staked_amount: new_stake.micro_ccd().try_into()?,
-                        })),
-                        DelegationEvent::DelegationStakeDecreased {
-                            delegator_id,
-                            new_stake,
-                        } => Ok(Event::DelegationStakeDecreased(DelegationStakeDecreased {
-                            delegator_id:      delegator_id.id.index.try_into()?,
-                            account_address:   details.sender.into(),
-                            new_staked_amount: new_stake.micro_ccd().try_into()?,
-                        })),
-                        DelegationEvent::DelegationSetRestakeEarnings {
-                            delegator_id,
-                            restake_earnings,
-                        } => {
-                            Ok(Event::DelegationSetRestakeEarnings(DelegationSetRestakeEarnings {
-                                delegator_id: delegator_id.id.index.try_into()?,
-                                account_address: details.sender.into(),
-                                restake_earnings,
-                            }))
-                        }
-                        DelegationEvent::DelegationSetDelegationTarget {
-                            delegator_id,
-                            delegation_target,
-                        } => Ok(Event::DelegationSetDelegationTarget(
-                            DelegationSetDelegationTarget {
-                                delegator_id:      delegator_id.id.index.try_into()?,
-                                account_address:   details.sender.into(),
-                                delegation_target: delegation_target.try_into()?,
-                            },
-                        )),
-                        DelegationEvent::DelegationAdded {
-                            delegator_id,
-                        } => Ok(Event::DelegationAdded(DelegationAdded {
-                            delegator_id:    delegator_id.id.index.try_into()?,
-                            account_address: details.sender.into(),
-                        })),
-                        DelegationEvent::DelegationRemoved {
-                            delegator_id,
-                        } => Ok(Event::DelegationRemoved(DelegationRemoved {
-                            delegator_id:    delegator_id.id.index.try_into()?,
-                            account_address: details.sender.into(),
-                        })),
-                        DelegationEvent::BakerRemoved {
-                            baker_id,
-                        } => Ok(Event::BakerRemoved(BakerRemoved {
-                            baker_id: baker_id.id.index.try_into()?,
-                        })),
-                    })
-                    .collect::<anyhow::Result<Vec<_>>>()?
-            }
-        },
-        BlockItemSummaryDetails::AccountCreation(details) => {
-            vec![Event::AccountCreated(AccountCreated {
-                account_address: details.address.into(),
-            })]
-        }
-        BlockItemSummaryDetails::Update(details) => {
-            vec![Event::ChainUpdateEnqueued(ChainUpdateEnqueued {
-                effective_time: DateTime::from_timestamp(
-                    details.effective_time.seconds.try_into()?,
-                    0,
-                )
-                .context("Failed to parse effective time")?,
-                payload:        true, // placeholder
-            })]
-        }
-    };
-    Ok(events)
 }
 
 impl TryFrom<concordium_rust_sdk::types::RejectReason> for TransactionRejectReason {
@@ -4984,386 +3982,6 @@ impl TryFrom<concordium_rust_sdk::types::RejectReason> for TransactionRejectReas
     }
 }
 
-impl From<concordium_rust_sdk::types::Memo> for TransferMemo {
-    fn from(value: concordium_rust_sdk::types::Memo) -> Self {
-        TransferMemo {
-            raw_hex: hex::encode(value.as_ref()),
-        }
-    }
-}
-
-#[derive(SimpleObject, serde::Serialize, serde::Deserialize)]
-pub struct Transferred {
-    amount: Amount,
-    from:   Address,
-    to:     AccountAddress,
-}
-
-#[derive(SimpleObject, serde::Serialize, serde::Deserialize)]
-pub struct AccountCreated {
-    account_address: AccountAddress,
-}
-
-#[derive(SimpleObject, serde::Serialize, serde::Deserialize)]
-pub struct AmountAddedByDecryption {
-    amount:          Amount,
-    account_address: AccountAddress,
-}
-
-#[derive(SimpleObject, serde::Serialize, serde::Deserialize)]
-#[graphql(complex)]
-pub struct BakerAdded {
-    staked_amount:    Amount,
-    restake_earnings: bool,
-    baker_id:         BakerId,
-    sign_key:         String,
-    election_key:     String,
-    aggregation_key:  String,
-}
-#[ComplexObject]
-impl BakerAdded {
-    async fn account_address<'a>(&self, _ctx: &Context<'a>) -> ApiResult<AccountAddress> {
-        todo_api!()
-    }
-}
-
-#[derive(SimpleObject, serde::Serialize, serde::Deserialize)]
-#[graphql(complex)]
-pub struct BakerKeysUpdated {
-    baker_id:        BakerId,
-    sign_key:        String,
-    election_key:    String,
-    aggregation_key: String,
-}
-#[ComplexObject]
-impl BakerKeysUpdated {
-    async fn account_address<'a>(&self, _ctx: &Context<'a>) -> ApiResult<AccountAddress> {
-        todo_api!()
-    }
-}
-
-#[derive(SimpleObject, serde::Serialize, serde::Deserialize)]
-#[graphql(complex)]
-pub struct BakerRemoved {
-    baker_id: BakerId,
-}
-#[ComplexObject]
-impl BakerRemoved {
-    async fn account_address<'a>(&self, _ctx: &Context<'a>) -> ApiResult<AccountAddress> {
-        todo_api!()
-    }
-}
-
-#[derive(SimpleObject, serde::Serialize, serde::Deserialize)]
-#[graphql(complex)]
-pub struct BakerSetRestakeEarnings {
-    baker_id:         BakerId,
-    restake_earnings: bool,
-}
-#[ComplexObject]
-impl BakerSetRestakeEarnings {
-    async fn account_address<'a>(&self, _ctx: &Context<'a>) -> ApiResult<AccountAddress> {
-        todo_api!()
-    }
-}
-
-#[derive(SimpleObject, serde::Serialize, serde::Deserialize)]
-#[graphql(complex)]
-pub struct BakerStakeDecreased {
-    baker_id:          BakerId,
-    new_staked_amount: Amount,
-}
-#[ComplexObject]
-impl BakerStakeDecreased {
-    async fn account_address<'a>(&self, _ctx: &Context<'a>) -> ApiResult<AccountAddress> {
-        todo_api!()
-    }
-}
-
-#[derive(SimpleObject, serde::Serialize, serde::Deserialize)]
-#[graphql(complex)]
-pub struct BakerStakeIncreased {
-    baker_id:          BakerId,
-    new_staked_amount: Amount,
-}
-#[ComplexObject]
-impl BakerStakeIncreased {
-    async fn account_address<'a>(&self, _ctx: &Context<'a>) -> ApiResult<AccountAddress> {
-        todo_api!()
-    }
-}
-
-#[derive(SimpleObject, serde::Serialize, serde::Deserialize)]
-#[graphql(complex)]
-pub struct ContractInitialized {
-    module_ref:        String,
-    contract_address:  ContractAddress,
-    amount:            Amount,
-    init_name:         String,
-    version:           ContractVersion,
-    // All logged events by the smart contract during the transaction execution.
-    contract_logs_raw: Vec<Vec<u8>>,
-}
-
-#[ComplexObject]
-impl ContractInitialized {
-    async fn events_as_hex(&self) -> ApiResult<connection::Connection<String, String>> {
-        let mut connection = connection::Connection::new(true, true);
-
-        self.contract_logs_raw.iter().enumerate().for_each(|(index, log)| {
-            connection.edges.push(connection::Edge::new(index.to_string(), hex::encode(log)));
-        });
-
-        // TODO: pagination info but not used at front-end currently (issue#318).
-
-        Ok(connection)
-    }
-
-    async fn events<'a>(
-        &self,
-        ctx: &Context<'a>,
-    ) -> ApiResult<connection::Connection<String, String>> {
-        let pool = get_pool(ctx)?;
-
-        let row = sqlx::query!(
-            "
-            SELECT
-                contracts.module_reference as module_reference,
-                name as contract_name,
-                schema as display_schema
-            FROM contracts
-            JOIN smart_contract_modules ON smart_contract_modules.module_reference = \
-             contracts.module_reference
-            WHERE index = $1 AND sub_index = $2
-            ",
-            self.contract_address.index.0 as i64,
-            self.contract_address.sub_index.0 as i64
-        )
-        .fetch_optional(pool)
-        .await?
-        .ok_or(ApiError::NotFound)?;
-
-        // Get the event schema if it exists.
-        let opt_event_schema = if let Some(event_schema) = row.display_schema.as_ref() {
-            let versioned_schema =
-                VersionedModuleSchema::new(event_schema, &None).map_err(|_| {
-                    ApiError::InternalError(
-                        "Database bytes should be a valid VersionedModuleSchema".to_string(),
-                    )
-                })?;
-
-            versioned_schema.get_event_schema(&row.contract_name).ok()
-        } else {
-            None
-        };
-
-        let mut connection = connection::Connection::new(true, true);
-
-        for (index, log) in self.contract_logs_raw.iter().enumerate() {
-            let decoded_log = decode_value_with_schema(
-                opt_event_schema.as_ref(),
-                log,
-                SmartContractSchemaNames::Event,
-            );
-
-            connection.edges.push(connection::Edge::new(index.to_string(), decoded_log));
-        }
-
-        // TODO: pagination info but not used at front-end currently (issue#318).
-
-        Ok(connection)
-    }
-}
-
-#[derive(Enum, Copy, Clone, PartialEq, Eq, serde::Serialize, serde::Deserialize)]
-pub enum ContractVersion {
-    V0,
-    V1,
-}
-
-impl From<concordium_rust_sdk::types::smart_contracts::WasmVersion> for ContractVersion {
-    fn from(value: concordium_rust_sdk::types::smart_contracts::WasmVersion) -> Self {
-        use concordium_rust_sdk::types::smart_contracts::WasmVersion;
-        match value {
-            WasmVersion::V0 => ContractVersion::V0,
-            WasmVersion::V1 => ContractVersion::V1,
-        }
-    }
-}
-
-#[derive(Debug, thiserror::Error, Clone)]
-#[error("Invalid contract version: {0}")]
-pub struct InvalidContractVersionError(i32);
-
-impl TryFrom<i32> for ContractVersion {
-    type Error = InvalidContractVersionError;
-
-    fn try_from(value: i32) -> Result<Self, Self::Error> {
-        match value {
-            0 => Ok(ContractVersion::V0),
-            1 => Ok(ContractVersion::V1),
-            _ => Err(InvalidContractVersionError(value)),
-        }
-    }
-}
-
-#[derive(SimpleObject, serde::Serialize, serde::Deserialize)]
-pub struct ContractModuleDeployed {
-    module_ref: String,
-}
-
-#[derive(SimpleObject, serde::Serialize, serde::Deserialize)]
-pub struct ContractCall {
-    contract_updated: ContractUpdated,
-}
-
-#[derive(SimpleObject, serde::Serialize, serde::Deserialize)]
-pub struct CredentialDeployed {
-    reg_id:          String,
-    account_address: AccountAddress,
-}
-
-#[derive(SimpleObject, serde::Serialize, serde::Deserialize)]
-pub struct CredentialKeysUpdated {
-    cred_id: String,
-}
-
-#[derive(SimpleObject, serde::Serialize, serde::Deserialize)]
-pub struct CredentialsUpdated {
-    account_address:  AccountAddress,
-    new_cred_ids:     Vec<String>,
-    removed_cred_ids: Vec<String>,
-    new_threshold:    Byte,
-}
-
-#[derive(SimpleObject, serde::Serialize, serde::Deserialize)]
-#[graphql(complex)]
-pub struct DataRegistered {
-    data_as_hex: String,
-}
-
-#[ComplexObject]
-impl DataRegistered {
-    async fn decoded(&self) -> ApiResult<DecodedText> {
-        let decoded_data = hex::decode(&self.data_as_hex).map_err(|e| {
-            error!("Invalid hex encoding {:?} in a controlled environment", e);
-            ApiError::InternalError("Failed to decode hex data".to_string())
-        })?;
-
-        Ok(DecodedText::from_bytes(decoded_data.as_slice()))
-    }
-}
-
-#[derive(SimpleObject, serde::Serialize, serde::Deserialize)]
-pub struct DecodedText {
-    text:        String,
-    decode_type: TextDecodeType,
-}
-
-impl DecodedText {
-    /// Attempt to parse the bytes as a CBOR string otherwise use HEX to present
-    /// the bytes.
-    fn from_bytes(bytes: &[u8]) -> Self {
-        if let Ok(text) = ciborium::from_reader::<String, _>(bytes) {
-            Self {
-                text,
-                decode_type: TextDecodeType::Cbor,
-            }
-        } else {
-            Self {
-                text:        hex::encode(bytes),
-                decode_type: TextDecodeType::Hex,
-            }
-        }
-    }
-}
-
-#[derive(Enum, Copy, Clone, PartialEq, Eq, serde::Serialize, serde::Deserialize)]
-pub enum TextDecodeType {
-    Cbor,
-    Hex,
-}
-
-#[derive(SimpleObject, serde::Serialize, serde::Deserialize)]
-pub struct EncryptedAmountsRemoved {
-    account_address:      AccountAddress,
-    new_encrypted_amount: String,
-    input_amount:         String,
-    up_to_index:          u64,
-}
-
-impl TryFrom<concordium_rust_sdk::types::EncryptedAmountRemovedEvent> for EncryptedAmountsRemoved {
-    type Error = anyhow::Error;
-
-    fn try_from(
-        removed: concordium_rust_sdk::types::EncryptedAmountRemovedEvent,
-    ) -> Result<Self, Self::Error> {
-        Ok(EncryptedAmountsRemoved {
-            account_address:      removed.account.into(),
-            new_encrypted_amount: serde_json::to_string(&removed.new_amount)?,
-            input_amount:         serde_json::to_string(&removed.input_amount)?,
-            up_to_index:          removed.up_to_index.index,
-        })
-    }
-}
-
-#[derive(SimpleObject, serde::Serialize, serde::Deserialize)]
-pub struct EncryptedSelfAmountAdded {
-    account_address:      AccountAddress,
-    new_encrypted_amount: String,
-    amount:               Amount,
-}
-
-#[derive(SimpleObject, serde::Serialize, serde::Deserialize)]
-pub struct NewEncryptedAmount {
-    account_address:  AccountAddress,
-    new_index:        u64,
-    encrypted_amount: String,
-}
-
-impl TryFrom<concordium_rust_sdk::types::NewEncryptedAmountEvent> for NewEncryptedAmount {
-    type Error = anyhow::Error;
-
-    fn try_from(
-        added: concordium_rust_sdk::types::NewEncryptedAmountEvent,
-    ) -> Result<Self, Self::Error> {
-        Ok(NewEncryptedAmount {
-            account_address:  added.receiver.into(),
-            new_index:        added.new_index.index,
-            encrypted_amount: serde_json::to_string(&added.encrypted_amount)?,
-        })
-    }
-}
-
-#[derive(SimpleObject, serde::Serialize, serde::Deserialize)]
-#[graphql(complex)]
-pub struct TransferMemo {
-    raw_hex: String,
-}
-
-#[ComplexObject]
-impl TransferMemo {
-    async fn decoded(&self) -> ApiResult<DecodedText> {
-        let decoded_data = hex::decode(&self.raw_hex).map_err(|e| {
-            error!("Invalid hex encoding {:?} in a controlled environment", e);
-            ApiError::InternalError("Failed to decode hex data".to_string())
-        })?;
-
-        Ok(DecodedText::from_bytes(decoded_data.as_slice()))
-    }
-}
-
-#[derive(SimpleObject, serde::Serialize, serde::Deserialize)]
-pub struct TransferredWithSchedule {
-    from_account_address: AccountAddress,
-    to_account_address:   AccountAddress,
-    total_amount:         Amount,
-    // TODO: amountsSchedule("Returns the first _n_ elements from the list." first: Int "Returns
-    // the elements in the list that come after the specified cursor." after: String "Returns the
-    // last _n_ elements from the list." last: Int "Returns the elements in the list that come
-    // before the specified cursor." before: String): AmountsScheduleConnection
-}
-
 #[derive(SimpleObject, serde::Serialize, serde::Deserialize)]
 #[graphql(complex)]
 pub struct ModuleReferenceEvent {
@@ -5554,329 +4172,6 @@ struct ModuleReferenceContractLinkEventsCollectionSegment {
     total_count: u64,
 }
 
-#[derive(SimpleObject, serde::Serialize, serde::Deserialize)]
-pub struct ChainUpdateEnqueued {
-    effective_time: DateTime,
-    // effective_immediately: bool, // Not sure this makes sense.
-    payload:        bool, // ChainUpdatePayload,
-}
-
-// union ChainUpdatePayload = MinBlockTimeUpdate | TimeoutParametersUpdate |
-// FinalizationCommitteeParametersUpdate | BlockEnergyLimitUpdate |
-// GasRewardsCpv2Update | ProtocolChainUpdatePayload |
-// ElectionDifficultyChainUpdatePayload | EuroPerEnergyChainUpdatePayload |
-// MicroCcdPerEuroChainUpdatePayload | FoundationAccountChainUpdatePayload |
-// MintDistributionChainUpdatePayload |
-// TransactionFeeDistributionChainUpdatePayload | GasRewardsChainUpdatePayload |
-// BakerStakeThresholdChainUpdatePayload | RootKeysChainUpdatePayload |
-// Level1KeysChainUpdatePayload | AddAnonymityRevokerChainUpdatePayload |
-// AddIdentityProviderChainUpdatePayload | CooldownParametersChainUpdatePayload
-// | PoolParametersChainUpdatePayload | TimeParametersChainUpdatePayload |
-// MintDistributionV1ChainUpdatePayload
-#[derive(SimpleObject, serde::Serialize, serde::Deserialize)]
-pub struct ChainUpdatePayload {
-    todo: bool,
-}
-
-#[derive(SimpleObject, serde::Serialize, serde::Deserialize)]
-pub struct ContractInterrupted {
-    contract_address:  ContractAddress,
-    // All logged events by the smart contract during this section of the transaction execution.
-    contract_logs_raw: Vec<Vec<u8>>,
-}
-
-#[ComplexObject]
-impl ContractInterrupted {
-    async fn events_as_hex(&self) -> ApiResult<connection::Connection<String, String>> {
-        let mut connection = connection::Connection::new(true, true);
-
-        self.contract_logs_raw.iter().enumerate().for_each(|(index, log)| {
-            connection.edges.push(connection::Edge::new(index.to_string(), hex::encode(log)));
-        });
-
-        // TODO: pagination info but not used at front-end currently (issue#318).
-
-        Ok(connection)
-    }
-
-    async fn events<'a>(
-        &self,
-        ctx: &Context<'a>,
-    ) -> ApiResult<connection::Connection<String, String>> {
-        let pool = get_pool(ctx)?;
-
-        let row = sqlx::query!(
-            "
-            SELECT
-                contracts.module_reference as module_reference,
-                name as contract_name,
-                schema as display_schema
-            FROM contracts
-            JOIN smart_contract_modules ON smart_contract_modules.module_reference = \
-             contracts.module_reference
-            WHERE index = $1 AND sub_index = $2
-            ",
-            self.contract_address.index.0 as i64,
-            self.contract_address.sub_index.0 as i64
-        )
-        .fetch_optional(pool)
-        .await?
-        .ok_or(ApiError::NotFound)?;
-
-        // Get the event schema if it exists.
-        let opt_event_schema = if let Some(event_schema) = row.display_schema.as_ref() {
-            let versioned_schema =
-                VersionedModuleSchema::new(event_schema, &None).map_err(|_| {
-                    ApiError::InternalError(
-                        "Database bytes should be a valid VersionedModuleSchema".to_string(),
-                    )
-                })?;
-
-            versioned_schema.get_event_schema(&row.contract_name).ok()
-        } else {
-            None
-        };
-
-        let mut connection = connection::Connection::new(true, true);
-
-        for (index, log) in self.contract_logs_raw.iter().enumerate() {
-            let decoded_log = decode_value_with_schema(
-                opt_event_schema.as_ref(),
-                log,
-                SmartContractSchemaNames::Event,
-            );
-
-            connection.edges.push(connection::Edge::new(index.to_string(), decoded_log));
-        }
-
-        // TODO: pagination info but not used at front-end currently (issue#318).
-
-        Ok(connection)
-    }
-}
-
-#[derive(SimpleObject, serde::Serialize, serde::Deserialize)]
-pub struct ContractResumed {
-    contract_address: ContractAddress,
-    success:          bool,
-}
-
-#[derive(SimpleObject, serde::Serialize, serde::Deserialize)]
-#[graphql(complex)]
-pub struct ContractUpdated {
-    contract_address:  ContractAddress,
-    instigator:        Address,
-    amount:            Amount,
-    receive_name:      String,
-    version:           ContractVersion,
-    // All logged events by the smart contract during this section of the transaction execution.
-    contract_logs_raw: Vec<Vec<u8>>,
-    input_parameter:   Vec<u8>,
-}
-
-#[ComplexObject]
-impl ContractUpdated {
-    async fn message_as_hex(&self) -> ApiResult<String> { Ok(hex::encode(&self.input_parameter)) }
-
-    async fn message<'a>(&self, ctx: &Context<'a>) -> ApiResult<String> {
-        let pool = get_pool(ctx)?;
-
-        let row = sqlx::query!(
-            "
-            SELECT
-                contracts.module_reference as module_reference,
-                name as contract_name,
-                schema as display_schema
-            FROM contracts
-            JOIN smart_contract_modules ON smart_contract_modules.module_reference = \
-             contracts.module_reference
-            WHERE index = $1 AND sub_index = $2
-            ",
-            self.contract_address.index.0 as i64,
-            self.contract_address.sub_index.0 as i64
-        )
-        .fetch_optional(pool)
-        .await?
-        .ok_or(ApiError::NotFound)?;
-
-        // Get the receive param schema if it exists.
-        let opt_receive_param_schema = if let Some(event_schema) = row.display_schema.as_ref() {
-            let versioned_schema =
-                VersionedModuleSchema::new(event_schema, &None).map_err(|_| {
-                    ApiError::InternalError(
-                        "Database bytes should be a valid VersionedModuleSchema".to_string(),
-                    )
-                })?;
-
-            versioned_schema
-                .get_receive_param_schema(
-                    &row.contract_name,
-                    ReceiveName::new_unchecked(&self.receive_name).entrypoint_name().into(),
-                )
-                .ok()
-        } else {
-            None
-        };
-
-        let decoded_input_parameter = decode_value_with_schema(
-            opt_receive_param_schema.as_ref(),
-            &self.input_parameter,
-            SmartContractSchemaNames::InputParameterReceiveFunction,
-        );
-
-        Ok(decoded_input_parameter)
-    }
-
-    async fn events_as_hex(&self) -> ApiResult<connection::Connection<String, String>> {
-        let mut connection = connection::Connection::new(true, true);
-
-        self.contract_logs_raw.iter().enumerate().for_each(|(index, log)| {
-            connection.edges.push(connection::Edge::new(index.to_string(), hex::encode(log)));
-        });
-
-        // TODO: pagination info but not used at front-end currently (issue#318).
-
-        Ok(connection)
-    }
-
-    async fn events<'a>(
-        &self,
-        ctx: &Context<'a>,
-    ) -> ApiResult<connection::Connection<String, String>> {
-        let pool = get_pool(ctx)?;
-
-        let row = sqlx::query!(
-            "
-            SELECT
-                contracts.module_reference as module_reference,
-                name as contract_name,
-                schema as display_schema
-            FROM contracts
-            JOIN smart_contract_modules ON smart_contract_modules.module_reference = \
-             contracts.module_reference
-            WHERE index = $1 AND sub_index = $2
-            ",
-            self.contract_address.index.0 as i64,
-            self.contract_address.sub_index.0 as i64
-        )
-        .fetch_optional(pool)
-        .await?
-        .ok_or(ApiError::NotFound)?;
-
-        // Get the event schema if it exists.
-        let opt_event_schema = if let Some(event_schema) = row.display_schema.as_ref() {
-            let versioned_schema =
-                VersionedModuleSchema::new(event_schema, &None).map_err(|_| {
-                    ApiError::InternalError(
-                        "Database bytes should be a valid VersionedModuleSchema".to_string(),
-                    )
-                })?;
-
-            versioned_schema.get_event_schema(&row.contract_name).ok()
-        } else {
-            None
-        };
-
-        let mut connection = connection::Connection::new(true, true);
-
-        for (index, log) in self.contract_logs_raw.iter().enumerate() {
-            let decoded_log = decode_value_with_schema(
-                opt_event_schema.as_ref(),
-                log,
-                SmartContractSchemaNames::Event,
-            );
-
-            connection.edges.push(connection::Edge::new(index.to_string(), decoded_log));
-        }
-
-        // TODO: pagination info but not used at front-end currently (issue#318).
-
-        Ok(connection)
-    }
-}
-
-#[derive(SimpleObject, serde::Serialize, serde::Deserialize)]
-pub struct ContractUpgraded {
-    contract_address: ContractAddress,
-    from:             String,
-    to:               String,
-}
-
-#[derive(SimpleObject, serde::Serialize, serde::Deserialize)]
-pub struct BakerSetBakingRewardCommission {
-    baker_id:                 BakerId,
-    account_address:          AccountAddress,
-    baking_reward_commission: Decimal,
-}
-
-#[derive(SimpleObject, serde::Serialize, serde::Deserialize)]
-pub struct BakerSetFinalizationRewardCommission {
-    baker_id: BakerId,
-    account_address: AccountAddress,
-    finalization_reward_commission: Decimal,
-}
-
-#[derive(SimpleObject, serde::Serialize, serde::Deserialize)]
-pub struct BakerSetTransactionFeeCommission {
-    baker_id:                   BakerId,
-    account_address:            AccountAddress,
-    transaction_fee_commission: Decimal,
-}
-
-#[derive(SimpleObject, serde::Serialize, serde::Deserialize)]
-pub struct BakerSetMetadataURL {
-    baker_id:        BakerId,
-    account_address: AccountAddress,
-    metadata_url:    String,
-}
-
-#[derive(SimpleObject, serde::Serialize, serde::Deserialize)]
-pub struct BakerSetOpenStatus {
-    baker_id:        BakerId,
-    account_address: AccountAddress,
-    open_status:     BakerPoolOpenStatus,
-}
-
-#[derive(SimpleObject, serde::Serialize, serde::Deserialize)]
-pub struct DelegationAdded {
-    delegator_id:    AccountIndex,
-    account_address: AccountAddress,
-}
-
-#[derive(SimpleObject, serde::Serialize, serde::Deserialize)]
-pub struct DelegationRemoved {
-    delegator_id:    AccountIndex,
-    account_address: AccountAddress,
-}
-
-#[derive(SimpleObject, serde::Serialize, serde::Deserialize)]
-pub struct DelegationSetDelegationTarget {
-    delegator_id:      AccountIndex,
-    account_address:   AccountAddress,
-    delegation_target: DelegationTarget,
-}
-
-#[derive(SimpleObject, serde::Serialize, serde::Deserialize)]
-pub struct DelegationSetRestakeEarnings {
-    delegator_id:     AccountIndex,
-    account_address:  AccountAddress,
-    restake_earnings: bool,
-}
-
-#[derive(SimpleObject, serde::Serialize, serde::Deserialize)]
-pub struct DelegationStakeDecreased {
-    delegator_id:      AccountIndex,
-    account_address:   AccountAddress,
-    new_staked_amount: Amount,
-}
-
-#[derive(SimpleObject, serde::Serialize, serde::Deserialize)]
-pub struct DelegationStakeIncreased {
-    delegator_id:      AccountIndex,
-    account_address:   AccountAddress,
-    new_staked_amount: Amount,
-}
-
 #[derive(SimpleObject)]
 struct BlockMetrics {
     /// The most recent block height. Equals the total length of the chain minus
@@ -5970,12 +4265,4 @@ impl MetricsPeriod {
             MetricsPeriod::LastYear => Duration::days(15),
         }
     }
-}
-
-#[derive(sqlx::Type, Copy, Clone)]
-#[sqlx(type_name = "transaction_type")] // only for PostgreSQL to match a type definition
-pub enum DbTransactionType {
-    Account,
-    CredentialDeployment,
-    Update,
 }
