+++ backend-rust/src/lib.rs
@@ -1,3 +1,9 @@
 pub mod graphql_api;
 pub mod indexer;
+
+pub mod address;
+pub mod decoded_text;
 pub mod metrics;
+pub mod transaction_event;
+pub mod transaction_type;
+pub mod types;
